requirements.txt
pytesseract
pillow
psycopg2
pdf2image
supabase
supabase-py
python-dotenv
flask
